# openapi-definition/openapi.yaml
---
openapi: 3.0.3
info:
  title: Baseball Management System
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local server
- url: http://192.168.0.119:8080
  description: A remote server
tags:
- name: BMS basic API
  description: BMS Common APIs for authenticated users
- name: BMS guest API
  description: API endpoints for guest users
- name: Mock BMS basic API
  description: Mock BMS Common APIs for authenticated users
paths:
  /api/v1/core/players:
    get:
      tags:
      - BMS basic API
      parameters:
      - name: limit
        in: query
        schema:
          default: 50
          type: number
      - name: offset
        in: query
        schema:
          type: number
      - name: team_ids
        in: query
        description: filter with team ids
        schema:
          type: array
          items:
            format: int32
            type: integer
        style: form
        explode: false
      responses:
        "200":
          description: OK
    post:
      tags:
      - BMS basic API
      responses:
        "200":
          description: OK
  /api/v1/core/players/{id}:
    get:
      tags:
      - BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
    put:
      tags:
      - BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
    delete:
      tags:
      - BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/v1/core/teams:
    get:
      tags:
      - BMS basic API
      responses:
        "200":
          description: A list of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamResponse"
              example:
              - id: 7935
                name: OLYBTSMAJN
                regulationAtBats: 3314.7160749257064
                createdAt: 2021-02-16T20:17:23.077965358
                updatedAt: 2023-02-16T20:17:23.077965358
                url_path: xdlpa
              - id: 7550
                name: MHTPJMX
                regulationAtBats: 9285.838092713186
                createdAt: 2020-02-25T22:05:54.233017228
                updatedAt: 2022-02-24T22:05:54.233017228
                url_path: oalzynjxvmosca
              - id: 1781
                name: HYZCGXM
                regulationAtBats: 8751.063662102344
                createdAt: 2017-09-14T04:09:42.712983081
                updatedAt: 2019-09-14T04:09:42.712983081
                url_path: viql
              - id: 813
                name: MROXVVJIS
                regulationAtBats: 2237.6594648617574
                createdAt: 2015-08-22T01:26:03.278734206
                updatedAt: 2017-08-21T01:26:03.278734206
                url_path: ulohy
        "500":
          description: Internal server error
    post:
      tags:
      - BMS basic API
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamRequest"
      responses:
        "200":
          description: OK
  /api/v1/core/teams/{id}:
    get:
      tags:
      - BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: The detail of a team
          content:
            application/json:
              schema:
                description: The response of a team
                required:
                - id
                - name
                - regulation_at_bats
                type: object
                properties:
                  id:
                    format: int32
                    type: integer
                  name:
                    type: string
                  url_path:
                    maxLength: 32
                    type: string
                  regulation_at_bats:
                    format: float
                    maximum: 5
                    minimum: 0
                    exclusiveMinimum: true
                    type: number
                    example: 1.5
                  created_at:
                    type: string
                    allOf:
                    - $ref: "#/components/schemas/LocalDateTime"
                  updated_at:
                    type: string
                    allOf:
                    - $ref: "#/components/schemas/LocalDateTime"
                  deleted_at:
                    type: string
                    allOf:
                    - $ref: "#/components/schemas/LocalDateTime"
              example:
                id: 123
                name: CJRDHNYW
                regulationAtBats: 9466.13924949258
                createdAt: 2020-12-11T11:14:01.474910848
                updatedAt: 2021-12-31T11:14:01.474910848
                url_path: gvoel
        "500":
          description: Internal server error
    put:
      tags:
      - BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamRequest"
      responses:
        "200":
          description: The result of updating a team
        "404":
          description: The team does not exist
        "500":
          description: Internal server error
    delete:
      tags:
      - BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The result of deleting a team
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: The team does not exist
        "500":
          description: Internal server error
  /api/v1/core/teams/{id}/players:
    get:
      tags:
      - BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: The detail of a team
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerResponse"
        "500":
          description: Internal server error
  /dummy/api/v1/core/games:
    get:
      tags:
      - Mock BMS basic API
      parameters:
      - name: limit
        in: query
        schema:
          default: 50
          minimum: 0
          type: number
      - name: offset
        in: query
        schema:
          minimum: 0
          type: number
      responses:
        "200":
          description: Game is found
        "404":
          description: Game is not found
  /dummy/api/v1/core/games/{id}:
    get:
      tags:
      - Mock BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          minimum: 0
          type: integer
      responses:
        "200":
          description: Game is found
        "404":
          description: Game is not found
  /dummy/api/v1/core/players:
    get:
      tags:
      - Mock BMS basic API
      parameters:
      - name: limit
        in: query
        schema:
          default: 50
          type: number
      - name: offset
        in: query
        schema:
          type: number
      - name: team_ids
        in: query
        description: filter with team ids
        schema:
          type: array
          items:
            format: int32
            type: integer
        style: form
        explode: false
      responses:
        "200":
          description: OK
    post:
      tags:
      - Mock BMS basic API
      responses:
        "200":
          description: OK
  /dummy/api/v1/core/players/{id}:
    get:
      tags:
      - Mock BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: OK
    put:
      tags:
      - Mock BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
    delete:
      tags:
      - Mock BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /dummy/api/v1/core/stadiums:
    get:
      tags:
      - Mock BMS basic API
      parameters:
      - name: limit
        in: query
        schema:
          default: 50
          minimum: 0
          type: number
      - name: offset
        in: query
        schema:
          minimum: 0
          type: number
      responses:
        "200":
          description: OK
    post:
      tags:
      - Mock BMS basic API
      responses:
        "200":
          description: OK
  /dummy/api/v1/core/stadiums/{id}:
    get:
      tags:
      - Mock BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          minimum: 0
          type: integer
      responses:
        "200":
          description: Stadium found
        "404":
          description: Stadium not found
    put:
      tags:
      - Mock BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
    delete:
      tags:
      - Mock BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /dummy/api/v1/core/teams:
    get:
      tags:
      - Mock BMS basic API
      responses:
        "200":
          description: A list of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamResponse"
              example:
              - id: 7935
                name: OLYBTSMAJN
                regulationAtBats: 3314.7160749257064
                createdAt: 2021-02-16T20:17:23.077965358
                updatedAt: 2023-02-16T20:17:23.077965358
                url_path: xdlpa
              - id: 7550
                name: MHTPJMX
                regulationAtBats: 9285.838092713186
                createdAt: 2020-02-25T22:05:54.233017228
                updatedAt: 2022-02-24T22:05:54.233017228
                url_path: oalzynjxvmosca
              - id: 1781
                name: HYZCGXM
                regulationAtBats: 8751.063662102344
                createdAt: 2017-09-14T04:09:42.712983081
                updatedAt: 2019-09-14T04:09:42.712983081
                url_path: viql
              - id: 813
                name: MROXVVJIS
                regulationAtBats: 2237.6594648617574
                createdAt: 2015-08-22T01:26:03.278734206
                updatedAt: 2017-08-21T01:26:03.278734206
                url_path: ulohy
        "500":
          description: Internal server error
    post:
      tags:
      - Mock BMS basic API
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamRequest"
      responses:
        "200":
          description: OK
  /dummy/api/v1/core/teams/{id}:
    get:
      tags:
      - Mock BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: The detail of a team
          content:
            application/json:
              schema:
                description: The response of a team
                required:
                - id
                - name
                - regulation_at_bats
                type: object
                properties:
                  id:
                    format: int32
                    type: integer
                  name:
                    type: string
                  url_path:
                    maxLength: 32
                    type: string
                  regulation_at_bats:
                    format: float
                    maximum: 5
                    minimum: 0
                    exclusiveMinimum: true
                    type: number
                    example: 1.5
                  created_at:
                    type: string
                    allOf:
                    - $ref: "#/components/schemas/LocalDateTime"
                  updated_at:
                    type: string
                    allOf:
                    - $ref: "#/components/schemas/LocalDateTime"
                  deleted_at:
                    type: string
                    allOf:
                    - $ref: "#/components/schemas/LocalDateTime"
              example:
                id: 123
                name: CJRDHNYW
                regulationAtBats: 9466.13924949258
                createdAt: 2020-12-11T11:14:01.474910848
                updatedAt: 2021-12-31T11:14:01.474910848
                url_path: gvoel
        "500":
          description: Internal server error
    put:
      tags:
      - Mock BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamRequest"
      responses:
        "200":
          description: The result of updating a team
        "404":
          description: The team does not exist
        "500":
          description: Internal server error
    delete:
      tags:
      - Mock BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The result of deleting a team
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: The team does not exist
        "500":
          description: Internal server error
  /dummy/api/v1/core/teams/{id}/players:
    get:
      tags:
      - Mock BMS basic API
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: The detail of a team
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerResponse"
        "500":
          description: Internal server error
components:
  schemas:
    LocalDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50
    PlayerResponse:
      type: object
      properties:
        id:
          format: int32
          type: integer
        name:
          type: string
        team_id:
          format: int32
          type: integer
        back_number:
          type: string
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
        updatedAt:
          $ref: "#/components/schemas/LocalDateTime"
        deletedAt:
          $ref: "#/components/schemas/LocalDateTime"
    TeamRequest:
      type: object
      properties:
        id:
          format: int32
          type: integer
        name:
          type: string
        url_path:
          maxLength: 32
          minLength: 4
          type: string
        regulationAtBats:
          format: double
          maximum: 5
          minimum: 0
          type: number
    TeamResponse:
      description: The response of a team
      required:
      - id
      - name
      - regulation_at_bats
      type: object
      properties:
        id:
          format: int32
          type: integer
        name:
          type: string
        url_path:
          maxLength: 32
          type: string
        regulation_at_bats:
          format: float
          maximum: 5
          minimum: 0
          exclusiveMinimum: true
          type: number
          example: 1.5
        created_at:
          type: string
          allOf:
          - $ref: "#/components/schemas/LocalDateTime"
        updated_at:
          type: string
          allOf:
          - $ref: "#/components/schemas/LocalDateTime"
        deleted_at:
          type: string
          allOf:
          - $ref: "#/components/schemas/LocalDateTime"